src/test/java/org/gd/leetcode/p0713/SolutionTest.java
package org.gd.leetcode.p0713;

import org.junit.jupiter.api.*;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.concurrent.TimeUnit;
import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.*;

/**
 * Test for {@link Solution}
 */
@DisplayName("LeetCode #713: Subarray Product Less Than K")
class SolutionTest {

    private static Stream<Arguments> args() {
        return Stream.of(
                Arguments.of(new int[]{10, 5, 2, 6}, 100, 8)
        );
    }

    @ParameterizedTest
    @MethodSource("args")
    @Timeout(value = 100, unit = TimeUnit.MILLISECONDS)
    void numSubarrayProductLessThanK(int[] nums, int k, int expected) {
        assertEquals(expected, new Solution().numSubarrayProductLessThanK(nums, k));
    }
}
